convert piecewise function tests to vitest
